% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnchoredInversionClient.R
\docType{package}
\name{AnchoredInversionClient}
\alias{AnchoredInversionClient}
\alias{AnchoredInversionClient-package}
\title{AnchoredInversionClient: HTTP client and example applications for Anchored Inversion.}
\description{
More details to come.
}
