% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darcy.R
\name{darcy}
\alias{darcy}
\title{Steady state 1-D groundwater flow equation, using Darcy's law}
\usage{
darcy(K, dx, source = rep(0, length(K)), b0_type = c("dirichlet",
  "neumann"), b0, b1_type = c("dirichlet", "neumann"), b1)
}
\arguments{
\item{K}{Hydraulic conductivity (m/s); 1-D vector.}

\item{dx}{Step size (m) of numerical grid.}

\item{source}{Volumetric source rate per unit volume;
\eqn{(m^3/s) / {m^3}}, ie \eqn{1/s}.}

\item{b0_type}{Type of upstream boundary condition.}

\item{b0}{Value of upstream boundary condition.
If \code{b0_type} is \code{'dirichlet'},
\code{b0} provides a head value (m).
If \code{b0_type} is \code{'neumann'},
\code{b0} provides water flux at boundary.
Flux is volumetric rate per volume; same unit as
\code{source}.}

\item{b1_type}{Type of downstream boundary condition.}

\item{b1}{Value of downstream boundary condition.}
}
\value{
Steady-state hydraulic head (m) in each grid cell;
     same length as \code{K}.
}
\description{
Solve the 1-D ODE:
\deqn{\frac{d(K \frac{dh}{dx})}{dx} = s}.
}
\details{
The model domain is discretized uniformly.
\code{K} (conductivity), \code{s}, \code{h} (head)
are all evaluated at the model grids.

The possible value range of \code{K} is huge (say 13 orders of magnitude).
Typical value range for good groundwater conditions is
\eqn{1e^{-5}} to \eqn{1e^{-1}} m/s.
}
\references{
Schwartz and Zhang, Fundamentals of Ground Water, Wiley, 2003.
}
