% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darcy.R
\name{darcy_field_transform}
\alias{darcy_field_transform}
\title{Transform the field value vector between natural and model scales}
\usage{
darcy_field_transform(x, reverse = FALSE)
}
\arguments{
\item{x}{The field value vector.}

\item{reverse}{If \code{FALSE} (the default), transform from natural scale to model scale;
If \code{TRUE} (that is, "reverse transformation"), transform from model scale to natural scale.}
}
\value{
Transformed field value vector.
}
\description{
Because Anchored Inversion works on a spatial variable that is defined by
the entire real line, i.e. \eqn{(-\infty, \infty)},
this transformation is needed if the actual field
is not defined on the real line. A typical case is that the actual field value
is positive. In this case, the log transformation may be suitable.
If the actual field value is already define on the real line,
then this transformation is not needed.
}
\details{
For the Darcy example application, we assume the field value is
on \eqn{[\text{K_xmin}, \infty)} (referred to as the "natural scale"),
and use a log transformation to map it onto the real line (referred to as the "model scale").

Field simulations requested from the Anchored Inversion server are always in the model scale.
Before running the forward model with these fields, the fields need to be reverse-transformed
into the natual scale. If the field simulations are to be used for other purposes,
they need to be reverse-transformed similarly.
}
