% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_fullxyz}
\alias{grid_fullxyz}
\title{The model grid is described by a list with three members:
  list(from = , by = , len = )
where
  'from' is coord of the _center_ of the first cell.
  'by' is step size between cells (ie, cell centers).
  'len' is number of cells in each dimension.}
\usage{
grid_fullxyz(grid)
}
\description{
The three elements are vectors of the same length;
their length is the spatial dimensionality of the model.
}
\details{
All dimensions should have the same length unit,
so that 'x = a' and 'y = a' means the same physical length
in the two directions.

Usually the first dimension is 'X', west-east,
the second is 'Y', south-north,
the third is 'Z', bottom-top.
}
