% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HalfDome}
\alias{HalfDome}
\title{2D elevation map near Half Dome, Yosemite National Park.}
\usage{
data(HalfDome)
}
\description{
This is a 2D elevation map near Half Dome in Yosemite National Park.
It can be used as synthetic 2D data.
}
\details{
The elevation data were downloaded from the National Map Seamless
Server, \url{http://seamless.usgs.gov}, in `Grid Float' format,
National Elevation Dataset (NED) 1 Arc Second,
and read in using function \code{gridfloat}.
}
\keyword{datasets}
