% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_transform}
\alias{log_transform}
\title{Log transformation.}
\usage{
log_transform(x, lower = 0, reverse = FALSE)
}
\arguments{
\item{x}{Numerical.}

\item{lower}{Numerical.}

\item{reverse}{Reverse transform?}
}
\value{
If \code{reverse} is \code{TRUE}, return
     \code{log(x - lower)}.
     If \code{reverse} is \code{FALSE}, return
     \code{lower + exp(x)}.
}
\description{
Transform a real number on \code{[lower, Inf)}
to the whole real line, or reversely.
}
