% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_transform.R
\name{logit_transform}
\alias{logit_transform}
\title{Logit transformation.}
\usage{
logit_transform(x, lower = 0, upper = 1, reverse = FALSE)
}
\arguments{
\item{x}{Numerical.}

\item{lower}{Vector.}

\item{upper}{Vector.
\code{lower} and \code{upper} must be numericals and such that
\code{lower < upper}.
This is not checked within this function.}

\item{reverse}{Reverse transform?}
}
\value{
If \code{reverse} is \code{TRUE}, return
     \code{(lower + upper * exp(x)) / (1 + exp(x))}.
     If \code{reverse} is \code{FALSE} (the default), return
     \code{log((x - lower) / (upper - x))}.
     The return could contain finite, infinite, or \code{NA} values.
}
\description{
Transform a real number on a interval to the whole real line,
or reversely.
}
