% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darcy.R
\name{make_darcy_field}
\alias{make_darcy_field}
\title{Construct a synthetic (fake) 1-D field of hydraulic conductivity}
\usage{
make_darcy_field(len = 200)
}
\arguments{
\item{len}{Length of the field to be constructed.}
}
\value{
A vector of the constructed synthetic field.
}
\description{
The constructed field is used as the "true field" in the example
Darcy application. Obviously this is only used in synthetic studies.
In real applications, we do not know the true field and the true field
is the ultimate target that we try to infer.
}
\details{
Internally, this function randomly selects a chunk from the
\code{\link{Denali}} dataset, performs transformations to render
the values in a reasonable range in the natural scale of hydraulic conductivity,
and finally revserse-transformed the field to the model scale.

The random pick of the data chunk is determined by the current random seed.
}
\seealso{
\code{\link{darcy_field_transform}}, \code{\link{Denali}}, \code{\link{K_xmin}}
}
