% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darcy.R
\name{make_darcy_linear_data}
\alias{make_darcy_linear_data}
\title{Construct synthetic linear data based on a synthetic conductivity field}
\usage{
make_darcy_linear_data(grid, field, n_linear)
}
\arguments{
\item{grid}{The model grid definition.}

\item{field}{The synthetic conductivity field vector, in model scale.
Almost always, this is the output of function \code{\link{make_darcy_field}}.}

\item{n_linear}{The number of linear data points to create.}
}
\value{
Constructed data, if \code{n_linear > 0}; or \code{NULL} if \code{n_linear} is 0.
}
\description{
This obviously is only needed in synthetic studies.
In real applications, linear data would combe from field measurements
rather than such a construction function.
}
\details{
Linear data is defined as linear functions
of the field on the numerical grid, such as direct
measurements. Such data may or may not be available
in actual applications.
In this Darcy example application, we simply directly pick
conductivity values at random locations in the field as linear data.
The randomness depends on the current random seed.
}
