% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_field_1d}
\alias{plot_field_1d}
\title{Plot a 1D field curve.}
\usage{
plot_field_1d(field, grid, xlab = "X", ylab = "Y", type = "l", ...)
}
\arguments{
\item{field}{A vector of the field values.}

\item{grid}{Grid definition. See `grid.R`.}

\item{...}{Additional paramters passed to `lattice::xyplot`.}
}
\description{
Plot a 1D field curve.
}
