% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_summary}
\alias{print_summary}
\title{Print a summary of an \code{anchorit} object.}
\usage{
print_summary(obj, digits = 3, ...)
}
\arguments{
\item{digits}{passed on to \code{print}.}

\item{\dots}{additional arguments to \code{print}.}
}
\description{
Print a summary of an \code{anchorit} object.
}
