% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darcy.R
\docType{topic}
\name{Darcy}
\alias{Darcy}
\title{Darcy groundwater hydrology example application}
\description{
Darcy groundwater hydrology example application
}
\seealso{
\code{\link{darcy}},
    \code{\link{K_xmin}},
    \code{\link{darcy_field_transform}},
    \code{\link{darcy_forward_transform}},
    \code{\link{make_darcy_field}},
    \code{\link{make_darcy_forward_function}},
    \code{\link{make_darcy_linear_data}}
}
