% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darcy.R
\name{darcy}
\alias{darcy}
\title{Steady state 1-D groundwater flow equation, using Darcy's law.}
\usage{
darcy(K, dx, source = rep(0, length(K)), b0_type = c("dirichlet",
  "neumann"), b0, b1_type = c("dirichlet", "neumann"), b1)
}
\arguments{
\item{K}{hydraulic conductivity (m/s).
The possible value range is huge (say 13 orders of magnitude).
Typical value range for good groundwater conditions:
1e-5---1e-1 m/s.}

\item{dx}{step size (m) of numerical grid.}

\item{source}{volumetric source rate per unit volume
(m^3/s / m^3, ie 1/s).}

\item{b0_type}{type of upstream boundary condition.}

\item{b0}{value of upstream boundary condition.
If \code{b0_type} is \code{'dirichlet'},
\code{b0} provides a head value (m).
If \code{b0_type} is \code{'neumann'},
\code{b0} provides water flux at boundary.
Flux is volumetric rate per volume; same unit as
\code{source}.}

\item{b1_type}{type of downstream boundary condition.}

\item{b1}{value of downstream boundary condition.

The model domain is discretized uniformly.
\code{K}, \code{s}, \code{h} are all evaluated at the model grids.}
}
\value{
Statedy-state hydraulic head (m) in each grid cell.
     Same length as \code{K}.
}
\description{
Solve the 1-D ODE: \code{d(K dh/dx) / dx = s}.
}
\references{
Schwartz and Zhang, Fundamentals of Ground Water, Wiley, 2003.
}
