% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_summary}
\alias{print_summary}
\title{Print the summary that is returned by `Session$summarize_project`.}
\usage{
print_summary(obj, digits = 3, ...)
}
\arguments{
\item{obj}{The output of `Session$summarize_project`.}

\item{digits}{passed on to \code{print}.}

\item{\dots}{additional arguments to \code{print}.}
}
\description{
Print the summary that is returned by `Session$summarize_project`.
}
